Rework the grouped model choice fields
I'm not sure exactly why I wrote these the way I did in the first place.
One way or another, it didn't seem to work when I wanted to use it for
primary tags in brockman. I've done a quick rewrite on this and it
appears to be working fine. Borrowed from our pal Vitor:

https://simpleisbetterthancomplex.com/tutorial/2019/01/02/how-to-implement-grouped-model-choice-field.html